Upgrade `docx` from v7 to v9; Add SVG image support
# Changes

## 1. Upgrade `docx` from `^7.3.0` to `^9.0.3`

Upgrading `docx` from `^7.3.0` to `^9.0.3` introduces a few minor breaking changes that affect `myst-cli` and `myst-to-docx`, all of which are addressed in this commit.

1. `IImageOptions` (the object passed to the `new ImageRun()` constructor) now requires a `type` property set to one of `"jpg" | "png" | "gif" | "bmp" | "svg"`.

2. If `type === "svg"`, `IImageOptions` also requires a `fallback` property containing a raster (bitmap) rendering of the SVG for [pre-2019 versions of Word](https://support.microsoft.com/en-us/office/edit-svg-images-in-microsoft-365-69f29d39-194a-4072-8c35-dbe5e7ea528c) and other applications that don't support SVG in OOXML.

3. The `IPropertiesOptions` interface is no longer exported by `docx`. However, since `myst-to-docx` only uses its `footnotes` property, which has a simple type of `Record<string, { readonly children: readonly Paragraph[] }>`, the type can simply be defined in `myst-to-docx`.

4. `AlignmentType` has changed from an `enum` to `const` object. Enforcing strict type checks on its values now requires a more verbose syntax: `let alignment: (typeof AlignmentType)[keyof typeof AlignmentType];`.

## 2. Migrate from `buffer-image-size` to `image-size`

`image-size` is actively maintained; `buffer-image-size` is not.

`image-size` supports both Node.js and browser usage.

## 3. Add `@resvg/resvg-js` to automatically rasterize SVG to PNG

`myst-to-docx` now automatically rasterizes SVG images to PNG as a `fallback` for older versions of Word that don't support SVG images.

`resvg-js` is an actively-maintained library that renders SVG to PNG _without the need for a headless browser_ like Playwright. It supports both Node.js and browser usage via WASM.

From the [project's README](https://github.com/thx/resvg-js):

> `resvg-js` is a high-performance SVG renderer and toolkit, powered by Rust based resvg, with Node.js backend using napi-rs, also a pure WebAssembly backend.
>
> Features:
>
> - Fast, safe and zero dependencies, with correct output.
> - Convert SVG to PNG, includes cropping, scaling and setting the background color.
> - No need for `node-gyp` and `postinstall`, the `.node` file has been compiled for you.
> - Cross-platform support, including Apple M Chips.
> - Support for running as native addons in Deno.

The two relevant ESBuild CLI invocations now include a `--loader:.node=binary` argument to inline the `.node` file in the compiled JavaScript as a `Uint8Array`, which should help avoid issues with bundlers.

## 4. `getImageDimensions()` now works in Node.js

`getImageDimensions()` has been updated to use the `image-size` library, which supports both Node.js and browser usage.